LookupCustomer improvements
Improved the way I'm picking up the job I've created - turns out this is
returned from the original Post in the Location header, so I'm using
this to check on job progress. Currently loops infinitely with a message
box, next step is to use the queue lookup output to end the loop at the
correct time.